requirements.txt
numpy==1.21.6
pandas==1.3.5
scipy==1.7.3
ipykernel==6.16.2
xlrd==2.0.1
numba==0.56.4
gmpy2==2.2.1
sympy==1.10.1
requests==2.31.0
cupy-cuda12x==11.6.0